% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_app_function.R
\name{init_app}
\alias{init_app}
\title{Initialize the app by setting path}
\usage{
init_app(work_dir, overwrite = FALSE)
}
\arguments{
\item{work_dir}{The working directory where MetaboliteAnnotator will store its data}

\item{overwrite}{Whether to overwrite existing data if present}
}
\description{
This function initializes the MetaboliteAnnotator app by setting up working directories
and checking for required files. It should be called before run_ui().
}
