% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_app_function.R
\name{load_all}
\alias{load_all}
\title{Load all components of the MetaboliteAnnotator package}
\usage{
load_all()
}
\description{
This function loads all components of the MetaboliteAnnotator package,
including resolving package conflicts and setting up resource paths.
It should be called before using the app.
}
