% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_app_function.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load and resolve package dependencies}
\usage{
load_dependencies()
}
\description{
This function loads all required packages for the MetaboliteAnnotator
and resolves potential conflicts between packages.
}
